moved AWK script and PDOS summation functions to their own module
