+++ data/data_collector.py
@@ -31,6 +31,7 @@ from typing import Any, Callable, Optional
 
 from ui.display_data import display_dft_info, display_atomic_states, display_wannier_info
 from ui.ui_helpers import prompt_input, print_warning, print_header, console
+from utils.external_tools import run_awk_script, run_sum_pdos
 from utils.file_parser import *
 from core.project_setup import initialize_project
 from core.input_handler import get_atomic_states
@@ -532,49 +533,6 @@ def collect_atomic_states_info(paths: Dict[str, List[str]],
 
     collector = CollectorFactory.create_atomic_states_info_collector()
     return collector.collect(config)
-
-
-def run_awk_script(number_of_atomic_states: int,
-                   fermi_energy: float,
-                   kpdos_output_dir: str,
-                   projbands_dir: str) -> None:
-    """
-    Execute the AWK script to generate projected bands data.
-
-    Args:
-        number_of_atomic_states (int): Number of atomic states
-        fermi_energy (float): Fermi energy value
-        kpdos_output_dir (str): KPDOS output path
-        projbands_dir (str): Path to generate the projbands file
-
-    Raises:
-        CalledProcessError: If the AWK script execution fails
-    """
-    print_info("Calculating projected bands...")
-
-    awk_command = (
-        f"awk -v firststate=1 "
-        f"-v laststate={number_of_atomic_states} "
-        f"-v ef={fermi_energy} "
-        f"-f utils/projwfc_to_bands.awk {kpdos_output_dir} > {projbands_dir}"
-    )
-
-    run(awk_command, shell=True, check=True, capture_output=True)
-
-
-def run_sum_pdos(atomic_projection: Tuple[str, str]) -> None:
-    """
-    Executes the Quantum ESPRESSO sumpdos.x script to get the desired PDOS files.
-
-    """
-
-    print_info(f"Summing the PDOS files for {atomic_projection[0]}-{atomic_projection[1]}")
-
-    sum_pdos_command = (f"sumpdos.x "
-                        f"*\({atomic_projection[0]}\)*\({atomic_projection[1]}*\) "
-                        f"> pdos_{atomic_projection[0]}_{atomic_projection[1]}.dat")
-
-    run(sum_pdos_command, shell=True, check=True, capture_output=True)
 
 
 def generate_projected_bands(paths: Dict[str, List[str]],
